refatoracao do método de conversão pra base64
